day21-ForEachLoop

+++ src/day20_Arrays/AverageNumbers.java
@@ -5,7 +5,7 @@ public class AverageNumbers {
 
         int[] numbers = {10,20,30,40,50,60};
 
-        int sum =0; // 10+20+30...
+        double sum =0; // 10+20+30...
 
         for (int i = 0; i < numbers.length; i++) {
                 sum +=numbers[i];
